Simplify Orange and Green instructions
The specificness of the instructions made it take a long time for the
app to read each combo (5+ seconds), which interrupts the flow of
training and becomes hard to remember. Allow the user to find their own
flow of using lead and rear sides instead, unless the combo is short and
trains a specific sequence (orange techniques 4 and 5, starting with
uppercut was the only one for now).

Also removed a dangling comma.